added test program and some test objectives
darcs-hash:20070823201709-c8de0-e8894aceee594f8bd61b5d13d85c7ae5d0692dfa.gz

+++ stogo/stogo.cc
@@ -30,7 +30,7 @@ int stogo_minimize(int n,
 		   objective_func fgrad, void *data,
 		   double *x, double *fmin,
 		   const double *l, const double *u,
-		   long int maxeval, long int maxtime)
+		   long int maxeval, double maxtime)
 {
   GlobalParams params;
